add: add confug to eslint
