+++ package.json
@@ -6,14 +6,20 @@
   "type": "module",
   "scripts": {
     "start": "ts-node ./index.ts",
-    "start:dev": "nodemon ./index.js"
+    "start:dev": "nodemon ./index.ts",
+    "lint": "eslint src/**/*.ts",
+    "format": "eslint src/**/*.ts --fix"
   },
   "author": "Andrei Auchynnikau",
   "keywords": [],
   "license": "ISC",
   "devDependencies": {
     "@types/node": "^20.11.17",
+    "@typescript-eslint/eslint-plugin": "^7.0.1",
+    "@typescript-eslint/parser": "^7.0.1",
+    "eslint": "^8.56.0",
     "nodemon": "^3.0.3",
+    "prettier": "^3.2.5",
     "ts-node": "^10.9.2",
     "typescript": "^5.3.3"
   },
