+++ tsconfig.json
@@ -24,5 +24,5 @@
     "target": "ES2020"
   },
   "include": ["src/**/*"],
-  "exclude": ["dist"]
+  "exclude": ["dist", "node_modules"]
 }
